# swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "TASKS",
        "description": "TASKS APIs"
    },
    "host": "localhost:8080",
    "basePath": "/",
...
    "schemes": [
        "http"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
...
        "/task/{id}": {
            "get": {
                "tags": [
                    "Task"
                ],
                "summary": "get Task details",
                "description": "get task details",
                "operationId": "details",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid operation"
                    },
                    "500": {
                        "description": "server error"
                    }
                },
                "x-swagger-router-controller": "taskController"
            },
            "put": {
                "tags": [
                    "Task"
                ],
                "summary": "update task",
                "description": "update task",
                "operationId": "update",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "update task",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/updateTask"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid operation"
                    },
                    "500": {
                        "description": "server error"
                    }
                },
                "x-swagger-router-controller": "taskController"
            },
            "delete": {
                "tags": [
                    "Task"
                ],
                "summary": "delete Task",
                "description": "delete task",
                "operationId": "delete",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid operation"
                    },
                    "500": {
                        "description": "server error"
                    }
                },
                "x-swagger-router-controller": "taskController"
            }
        }
    },
    "definitions": {
...
        "updateTask": {
            "properties": {
                "title": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "priority": {
                    "type": "string",
                    "enum": [
                        "HIGH",
                        "MEDIUM",
                        "LOW"
                    ]
                },
                "status": {
                    "type": "string",
                    "enum": [
                        "PENDING",
                        "COMPLETED",
                        "INPROGRESS",
                        "MISSED"
                    ]
                }
            },
            "example": {
                "status": "COMPLETED"
            }
        }
    }
}